Add files via upload
The credit card approval system is an efficient and secure client-server-based application designed to facilitate the process of approving credit card requests from clients. Its primary purpose is to evaluate the eligibility of individuals who apply for a credit card and determine whether they meet the predefined criteria. Once the credit card application is approved, the client gains the ability to make transactions anywhere using the credit card. When a client initiates a transaction, the client's credit card details, transaction amount, and other relevant information are securely transmitted to the server, which is managed by the server.

The Technologies used are - Java and MySQL